Page has more responsibilities
Added the necessary logic for handling link mapping and verification,
but I need to implement more tests to ensure that everything is working
correctly. Some of the tests are also parameterized, so that I can add
more errors that I detect while running the program to them as time goes
on.